requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
redis
rq
python-dotenv
pydantic


# dev/test extras (optional)
pytest
httpx